can now show data on edit.php but not fully fixed
fixed the link to save the data in add.php. changed the code in try.php for the code for getting the data for the edit.php. fixed the data in gen.php.  on loc.php added select option for the province but set it to read only can now show the data on the selected crop on the list. added the code to show the data for a select field in edit.php. transfered the script for crops and location table on crop.php.